fix: fix types
