+++ packages/kit/src/routes/Tab/Home/type.ts
@@ -1,0 +1,7 @@
+export enum ETabHomeRoutes {
+  TabHome = 'TabHome',
+}
+
+export type ITabHomeParamList = {
+  [ETabHomeRoutes.TabHome]: undefined;
+};
