+++ packages/kit/src/routes/Tab/Me/TabMe.tsx
@@ -10,7 +10,7 @@ import useAppNavigation from '../../../hooks/useAppNavigation';
 import { EModalRoutes } from '../../Modal/type';
 import { ETabRoutes } from '../type';
 
-import type { ITabMeParamList } from './Routes';
+import type { ITabMeParamList } from './type';
 
 const TabMe = () => {
   const intl = useIntl();
