+++ packages/kit/src/routes/Tab/Swap/router.ts
@@ -1,0 +1,14 @@
+import type { ITabSubNavigatorConfig } from '@onekeyhq/components';
+
+import Swap from '../../../views/Swap';
+
+import { ETabSwapRoutes } from './type';
+
+export const swapRouters: ITabSubNavigatorConfig<any, any>[] = [
+  {
+    name: ETabSwapRoutes.TabSwap,
+    component: Swap,
+    rewrite: '/',
+    translationId: 'title__swap',
+  },
+];
