+++ packages/kit/src/routes/config/allowList.ts
@@ -3,7 +3,7 @@ import platformEnv from '@onekeyhq/shared/src/platformEnv';
 import { ERootRoutes } from '../enum';
 import { EGalleryRoutes } from '../Tab/Developer/Gallery/routes';
 import { ETabHomeRoutes } from '../Tab/Home/router';
-import { ETabSwapRoutes } from '../Tab/Swap/Routes';
+import { ETabSwapRoutes } from '../Tab/Swap/type';
 import { ETabRoutes } from '../Tab/type';
 
 interface IAllowSettingItem {
